requirements.txt

torch==2.0.0
torchvision==0.15.1
Pillow==9.5.0
